remove rubymine configs
