+++ lesson_1/objTreeResult.txt
@@ -1,0 +1,6 @@
+|1
+|  |__2
+|    |____3
+|    |____4
+|  |__5
+|    |____6
